// solar-proposal-app/src/generated/prisma/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    company: Prisma.$CompanyPayload<ExtArgs> | null
    leads: Prisma.$LeadPayload<ExtArgs>[]
    proposals: Prisma.$ProposalPayload<ExtArgs>[]
    scanJobs: Prisma.$ScanJobPayload<ExtArgs>[]
    campaigns: Prisma.$OutreachCampaignPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    email: string
    password: string
    name: string | null
    role: $Enums.Role
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["user"]>
  composites: {}
}