// solar-proposal-app/src/generated/prisma/models/Company.ts
export type CompanyModel = runtime.Types.Result.DefaultSelection<Prisma.$CompanyPayload>
// ...
export type $CompanyPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Company"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    logo: string | null
    website: string | null
    phone: string | null
    address: string | null
    primaryColor: string
    secondaryColor: string
    defaultElectricityRate: number
    defaultPanelWattage: number
    defaultPanelCostPerWatt: number
    defaultIncentivePercent: number
    attomApiKey: string | null
    resendApiKey: string | null
    fromEmail: string | null
    fromName: string | null
    calendlyLink: string | null
    userId: string
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["company"]>
  composites: {}
}